add pipes from Extra Buildcraft Pipes
blakmajik:
I am currently porting to 2.2, they should be out this weekend. But yes, there is no license on this mod, as there was none when I took it over. You make take any part of it and do as you wish :D

leftler:
Go ahead and take it, sorry it took so long to respond, I am no longer mataining those mods anyway.